Move config defaulting to lib/init.js
